+++ Editor/Aseprite/Chunks/CellExtra.cs
@@ -6,12 +6,7 @@ namespace UnityEditor.U2D.Aseprite
     internal class CellExtra : BaseChunk
     {
         public override ChunkTypes chunkType => ChunkTypes.CellExtra;
-
         public CellExtra(uint chunkSize) : base(chunkSize) { }
-
-        public override void Read(BinaryReader reader)
-        {
-            var bytes = reader.ReadBytes((int)m_ChunkSize - ChunkHeader.stride);
-        }
+        protected override void InternalRead(BinaryReader reader) { }
     }
 }