+++ Editor/Aseprite/Chunks/PaletteChunk.cs
@@ -20,7 +20,7 @@ namespace UnityEditor.U2D.Aseprite
         
         public PaletteChunk(uint chunkSize) : base(chunkSize) { }
         
-        public override void Read(BinaryReader reader)
+        protected override void InternalRead(BinaryReader reader)
         {
             noOfEntries = reader.ReadUInt32();
             firstColorIndex = reader.ReadUInt32();